fix: resolve 12 QA audit issues (2 HIGH, 5 MEDIUM, 3 LOW + 2 verified OK)
- Chain slots now show document ref numbers instead of "Not yet uploaded"
- InlineSearch suppressed on /search page to prevent dropdown trap
- Advanced search po_no fallback queries purchase_orders table directly
- Search page gets submit button with proper input padding
- Close order preserves chain_status mismatch with warning in response
- Email validation via EmailStr on customer create/update schemas
- completeness_pct standardized to int across schema and API
- Auto-derive order scenario from document evidence when UNKNOWN
- React Router v7 future flags added to suppress console warnings
- PATCH /so-number returns full POResponse with customer fields
- Blank ref_number results filtered from search output
- pydantic[email] added to requirements.txt

+++ backend/app/schemas/purchase_order.py
@@ -83,7 +83,7 @@ class ChainSlot(BaseModel):
 class ChainStatusResponse(BaseModel):
     po_id: UUID
     po_number: str
-    completeness_pct: float
+    completeness_pct: int
     chain: dict[str, list[ChainSlot]]
